move plot_utils for kwargs/options
