+++ magnetopy/experiments/zfcfc.py
@@ -8,14 +8,20 @@ import pandas as pd
 import matplotlib.pyplot as plt
 
 from magnetopy.data_files import DatFile, filename_label, plot_raw, plot_raw_residual
-from magnetopy.experiments.plot_utils import get_ylabel, handle_kwargs, handle_options
+from magnetopy.experiments.plot_utils import get_ylabel
 from magnetopy.experiments.utils import (
     add_uncorrected_moment_columns,
     num_digits_after_decimal,
     scale_dc_data,
 )
 from magnetopy.parsing_utils import find_temp_turnaround_point
-from magnetopy.plot_utils import default_colors, force_aspect, linear_color_gradient
+from magnetopy.plot_utils import (
+    default_colors,
+    force_aspect,
+    linear_color_gradient,
+    handle_kwargs,
+    handle_options,
+)
 
 
 class FieldDetectionError(Exception):
